pyproject.toml
[project]
name = "workshop-framework"
description = "Workshop Framework"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "requests>=2.31.0",
    "pyyaml>=6.0.1",
]

[project.scripts]
devx = "devx.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["devx"]
exclude = ["templates"]

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.pylance]
reportGeneralTypeIssues = "warning"
reportMissingImports = "warning"
reportMissingTypeStubs = "warning"
lineLength = 120

[tool.pylint.messages_control]
max-line-length = 120

[tool.brev]
instance_type = "g4dn.xlarge"
workspace_group_id = "default"
ports = [
    { port = 8888, name = "Jupyter" },
    { port = 3000, name = "VSCode" }
]
